requirements.txt
geobuf
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
Django>=2.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-simple-captcha==0.5.5
djangorestframework>=3.9.4
djangorestframework-gis>=0.14
djangorestframework-jwt==1.10.0
html5lib==0.999999999
idna==2.5
Markdown==2.6.8
Pillow==3.1.0
psycopg2>=2.7.1
pycparser==2.18
PyJWT==1.5.0
python-mimeparse==1.6.0
pytz==2017.2
requests
six==1.10.0
webencodings==0.5.1
django-redis
